+++ docs/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 Unreleased
 ^^^^^^^^^^
 - Added CaptionReadTimingError for invalid SCC timestamps
+- Added CaptionReadSyntaxError for invalid colors in SAMIReader
+- Raise CaptionReadTimingError when missing 'begin' or 'end' and 'dur' time on DFXPReader
 
 2.0.4
 ^^^^^